Tornat a fer fins on tenia, pero ara a la taula de historials es guarden els noms

+++ app/Http/Controllers/DeleteUserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class DeleteUserController extends Controller
 {
@@ -12,15 +13,20 @@ class DeleteUserController extends Controller
         $userId = $request->input('user_id');
         $user = User::find($userId);
 
-        if (auth()->user()->name === 'Professorat') {
-            // Your logic to delete the user should go here
-            $user->delete();
+        //Revisar que la $request->input('password') sigui la contrasenya de l'usuari autenticat
+        if (auth()->check() && Hash::check($request->input('password'), auth()->user()->password)) {
 
-            // Redirect or return a response as needed
-            return redirect()->route('gestionar-alumnes')->with('success', 'Alumne eliminat correctament.');
+            if (auth()->user()->name === 'Professorat') {
+                // Your logic to delete the user should go here
+                $user->delete();
+
+                // Redirect or return a response as needed
+                return redirect()->route('gestionar-alumnes')->with('success', 'Alumne eliminat correctament.');
+            }
+
+            // If the user does not have the necessary permissions, handle accordingly
+            return redirect()->route('gestionar-alumnes')->with('error', 'No tens permisos per eliminar aquest alumne.');
         }
-
-        // If the user does not have the necessary permissions, handle accordingly
-        return redirect()->route('gestionar-alumnes')->with('error', 'No tens permisos per eliminar aquest alumne.');
+        return redirect()->route('gestionar-alumnes')->with('error', 'Contrasenya incorrecta o usuari no autenticat.');
     }
 }
